apply to all atomic widgets
