--- assets/dev/js/editor/elements/types/atomic-heading.js
@@ -1,17 +1,0 @@
-import Base from '../types/base/element-base';
-import Model from 'elementor-elements/models/widget';
-import { default as View } from 'elementor-elements/views/atomic-heading';
-
-export default class AtomicHeading extends Base {
-	getType() {
-		return 'a-heading';
-	}
-
-	getView() {
-		return View;
-	}
-
-	getModel() {
-		return Model;
-	}
-}
