+++ assets/dev/js/editor/elements/types/atomic-widgets/atomic-image.js
@@ -1,0 +1,7 @@
+import AtomicWidgetBase from "../atomic-widget-base";
+
+export default class AtomicImage extends AtomicWidgetBase {
+	getType() {
+		return 'a-image';
+	}
+}
