+++ assets/dev/js/editor/elements/types/atomic-widgets/index.js
@@ -1,0 +1,2 @@
+export { default as AtomicHeading } from './atomic-heading';
+export { default as AtomicImage } from './atomic-image';
