+++ assets/dev/js/editor/elements/views/atomic-widget-base.js
@@ -1,0 +1,16 @@
+const WidgetView = require( 'elementor-elements/views/widget' );
+
+class AtomicWidgetBaseView extends WidgetView {
+	tagName() {
+		return 'div';
+	}
+
+	className() {
+		const atomicClasses = [ 'elementor-widget-v2', `elementor-widget-v2-${ this.model.id }` ];
+		const widgetClasses = ( this.model.getSetting( 'classes' )?.value || [] );
+
+		return atomicClasses.concat( widgetClasses ).join( ' ' );
+	}
+}
+
+module.exports = AtomicWidgetBaseView;
